change code in tables.py

+++ components/tables/tables.py
@@ -7,7 +7,7 @@ class Tables:
     def create_users_table(self):
         query = '''
         CREATE TABLE IF NOT EXISTS users (
-        ID INTEGER PRIMARY KEY,
+        ID BIGSERIAL PRIMARY KEY,
         FIRST_NAME VARCHAR(255) NOT NULL,
         LAST_NAME VARCHAR(255) NOT NULL,
         EMAIL VARCHAR(255) NOT NULL,
